requirements.txt
requests
langchain
langgraph
fastapi
uvicorn
python-dotenv
langchain-openrouter
